# Improve performance of restxml_unmarshal